requirements.txt
Flask==2.2.2
flask-restx==1.0.3
Flask-Cors==3.0.10
Flask-Log-Request-ID==0.10.1
gunicorn==20.1.0
PyMySQL==1.0.2
PyJWT==2.6.0
elasticsearch==8.5.3
requests==2.28.1
boto3==1.26.45
twilio==7.16.1
six==1.16.0
Werkzeug==2.2.2
pytz==2022.6
APScheduler==3.10.1
